requirements.txt
crewai
langchain
langchain_community
google_generativeai
langchain_google_genai
python_dotenv